Rename num variable to num_replicates_minus_samples
Make it clearer what the variable contains. The idea is detect if
any biological replicates were used at all.
If there are 2 repl. + 3 repl. then the calculation is 2 + 3 - 2 = 3
and because 3 > 1 then we let edgeR use the replicates and do not
force --dispersion argument.
Likewise, 4 replicates + 5 replicates - 2 samples = 7.

It might be preferred to require the number of replicates per
each sample to be > 1 but we assume the downstream code in edgeR is
smarter then us.