fix problem with numpy version
There was a problem with numpy version:
- we require numpy > 1.19.2
- tmhmm requires numpy < 1.24.0
because it uses feature which was removed starting from 1.24 ( it gives `AttributeError: module 'numpy' has no attribute 'int'`)

I tried to specify the version as `numpy>1.19.2,<1.24.0` but this lead to another error: `ERROR: Failed building wheel for numpy`
I specified version 1.22.3 as suggested here https://stackoverflow.com/questions/70565965/error-failed-building-wheel-for-numpy-error-could-not-build-wheels-for-numpy#comment126799390_70566445 and it fixed the problem. This numpy version requires python >= 3.8 so i also changed the version requirement for python.